+++ CHANGES.rst
@@ -17,6 +17,7 @@ Development
 - Turn off interrupts while powering down the HX711
 - Improve inline documentation and logging
 - Improve IRQ handling when reading the HX711
+- Sleep for 80 microseconds after pulling HX711 clock pin to high
 
 
 2019-06-17 0.4.0